Update debug tools
