+++ crates/prover/src/debug_tools/mod.rs
@@ -1,0 +1,2 @@
+pub mod assert_constraints;
+pub mod relation_tracker;
