+++ crates/prover/tests/prover.rs
@@ -1,10 +1,12 @@
+//! Run these tests with feature `relation-tracker` to see the relation tracker output.
 use std::collections::HashMap;
 use std::fs;
 
 use cairo_m_compiler::{compile_cairo, CompilerOptions};
 use cairo_m_prover::adapter::memory::Memory;
 use cairo_m_prover::adapter::{import_from_runner_output, Instructions, ProverInput};
-use cairo_m_prover::prover::{assert_constraints, prove_cairo_m};
+use cairo_m_prover::debug_tools::assert_constraints::assert_constraints;
+use cairo_m_prover::prover::prove_cairo_m;
 use cairo_m_prover::verifier::verify_cairo_m;
 use cairo_m_runner::run_cairo_program;
 use stwo_prover::core::fields::m31::M31;
@@ -62,8 +64,6 @@ fn test_prove_and_verify_empty_memory() {
     result.unwrap();
 }
 
-// Run this test with `cargo test -p cairo-m-prover test_prove_and_verify_fibonacci_program --release --features relation-tracker -- --nocapture`
-// to see the relation tracker output.
 #[test]
 fn test_prove_and_verify_fibonacci_program() {
     let source_path = format!(
@@ -125,7 +125,7 @@ fn test_all_constraints() {
     let source_path = format!(
         "{}/tests/test_data/{}",
         env!("CARGO_MANIFEST_DIR"),
-        "recursive_fibonacci.cm"
+        "fibonacci.cm"
     );
     let compiled_fib = compile_cairo(
         fs::read_to_string(&source_path).unwrap(),
